Consultar Laboratorios

+++ DTO/src/main/java/DTOLabComputo/centroLabDTO.java
@@ -8,18 +8,18 @@ import java.sql.Time;
  */
 public class centroLabDTO {
 
-    private int CentroLab_ID;
+    private long CentroLab_ID;
     private String Nombre;
     private Time HoraInicio;
     private Time HoraFin;
     private String Campus;
     private boolean estEliminado;
 
-    public int getCentroLab_ID() {
+    public long getCentroLab_ID() {
         return CentroLab_ID;
     }
 
-    public void setCentroLab_ID(int CentroLab_ID) {
+    public void setCentroLab_ID(long CentroLab_ID) {
         this.CentroLab_ID = CentroLab_ID;
     }
